Add docs and format.
