+++ lib/guardian/plug/pipeline.ex
@@ -78,19 +78,7 @@ if Code.ensure_loaded?(Plug) do
 
     When using plugs, you'll need to specify an error handler module
 
-    The error_handler module requires an `auth_error` function that receives the conn,
-    the reason tuple and the options
-
-    ```elixir
-    defmodule MyApp.AuthErrorHandler do
-      @behaviour Guardian.ErrorHandler
-
-      @impl Guardian.ErrorHandler
-      def auth_error(conn, {type, reason}, opts) do
-        ...
-      end
-    end
-    ```
+    See `Guardian.ErrorHandler` documentation for more details.
 
     ### Inline pipelines
